VGG Net with ADAM & SGD
This project presents the implementation of the VGG (Visual Geometry Group) convolutional neural network architecture in combination with two powerful optimization algorithms, ADAM and SGD (Stochastic Gradient Descent). The VGG architecture is known for its remarkable performance in image recognition tasks, making it a compelling choice for deep learning enthusiasts and researchers. By incorporating both ADAM and SGD, we offer a unique opportunity to compare the effects of these optimization methods on model training. This project is an invaluable resource for those looking to gain a deeper understanding of optimization algorithms and their impact on the training process, as well as for developers seeking to fine-tune their convolutional neural networks for various image-based tasks. Dive into the world of deep learning, VGG networks, and optimization strategies with "VGG Net with ADAM & SGD" to enhance your machine learning skills and model performance.